Create get-seed-node-ips.bash
This script can be used in an auto-scaling group to grab two IPs from each availability zone in an autoscaling group.  

It can be used to populate the seed nodes cassandra configuration file on startup.  This can be a building block to creating a C* cluster in an ASG without using a reflector, as the nodes will not need to use any external source for their seed IPs.  

The configuration file could be kept up to date with a cron job using this script to pull current seed nodes if a seed node is replaced because of failure.